+++ validators/topology_snapshot.go
@@ -13,6 +13,7 @@ import (
 	"code.vegaprotocol.io/vega/libs/crypto"
 	"code.vegaprotocol.io/vega/logging"
 	"code.vegaprotocol.io/vega/types"
+	"code.vegaprotocol.io/vega/types/num"
 
 	"github.com/golang/protobuf/proto"
 	tmtypes "github.com/tendermint/tendermint/abci/types"
@@ -73,6 +74,7 @@ func (t *Topology) serialiseNodes() []*snapshot.ValidatorState {
 					ExpectedNextHashSince: node.heartbeatTracker.expectedNexthashSince.UnixNano(),
 				},
 				ValidatorPower: node.validatorPower,
+				RankingScore:   node.rankingScore,
 			},
 		)
 	}
@@ -192,6 +194,7 @@ func (t *Topology) restore(ctx context.Context, topology *types.Topology) error 
 
 	vUpdates := []tmtypes.ValidatorUpdate{}
 
+	epochSeq := num.NewUint(t.epochSeq).String()
 	for _, node := range topology.ValidatorData {
 		t.log.Debug("restoring validator data snapshot", logging.String("nodeid", node.ValidatorUpdate.NodeId))
 		vs := &valState{
@@ -218,6 +221,7 @@ func (t *Topology) restore(ctx context.Context, topology *types.Topology) error 
 				expectedNexthashSince: time.Unix(0, node.HeartbeatTracker.ExpectedNextHashSince),
 			},
 			validatorPower: node.ValidatorPower,
+			rankingScore:   node.RankingScore,
 		}
 		for i := 0; i < 10; i++ {
 			vs.heartbeatTracker.blockSigs[i] = node.HeartbeatTracker.BlockSigs[i]
@@ -225,6 +229,11 @@ func (t *Topology) restore(ctx context.Context, topology *types.Topology) error 
 		t.validators[node.ValidatorUpdate.NodeId] = vs
 
 		t.sendValidatorUpdateEvent(ctx, vs.data, true)
+
+		// send an event with the current ranking of the validator
+		if node.RankingScore != nil {
+			t.broker.Send(events.NewValidatorRanking(ctx, epochSeq, node.ValidatorUpdate.NodeId, node.RankingScore.StakeScore, node.RankingScore.PerformanceScore, node.RankingScore.RankingScore, protoStatusToString(node.RankingScore.PreviousStatus), protoStatusToString(node.RankingScore.Status), int(node.RankingScore.VotingPower)))
+		}
 
 		// this node is started and expect to be a validator
 		// but so far we haven't seen ourselve as validators for